Adding code for getting nth fibonacci number in log n

+++ Algorithms/Fibonacci-In-Logn/src/power-in-logn.hpp
@@ -2,7 +2,7 @@
 #include <vector>
 using namespace std;
 
-double power(int x, int n) {
+double power(double x, int n) {
     int nTemp = n;
     double ans = 1;
     bool toInvert;
